+++ src/main/java/gift/controller/ProductController.java
@@ -1,27 +1,22 @@
 package gift.controller;
 
-import gift.domain.Category;
 import gift.domain.Product;
-import gift.error.NotFoundException;
 import gift.service.CategoryService;
 import gift.service.ProductService;
-import io.swagger.v3.oas.annotations.Operation;
-import io.swagger.v3.oas.annotations.tags.Tag;
 import jakarta.validation.Valid;
-import java.util.ArrayList;
-import java.util.List;
-import org.springframework.data.domain.Page;
-import org.springframework.data.domain.PageRequest;
-import org.springframework.data.domain.Pageable;
 import org.springframework.http.HttpStatus;
-import org.springframework.stereotype.Controller;
-import org.springframework.ui.Model;
-import org.springframework.web.bind.annotation.*;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.DeleteMapping;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.PutMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
 
-
-@Controller
-@RequestMapping("/products")
-@Tag(name = "Product", description = "상품 API")
+@RestController
+@RequestMapping("/api/products")
 public class ProductController {
 
     private final ProductService productService;
@@ -32,91 +27,32 @@ public class ProductController {
         this.categoryService = categoryService;
     }
 
-    //상품 전체 조회 페이지
-    @GetMapping
-    @ResponseStatus(HttpStatus.OK)
-    @Operation(summary = "전체 상품 조회", description = "상품 리스트를 페이지네이션 방식으로 출력합니다.")
-    public String showProductList(
-        @RequestParam(name = "page", required = false, defaultValue = "0") Integer page,
-        @RequestParam(name = "size", required = false, defaultValue = "10") Integer size,
-        Model model) {
-
-        Pageable pageable = PageRequest.of(page, size);
-        Page<Product> productsPage = productService.getAllProducts(pageable);
-
-        model.addAttribute("products", productsPage.getContent());
-        model.addAttribute("currentPage", productsPage.getNumber());
-        model.addAttribute("totalPages", productsPage.getTotalPages());
-        model.addAttribute("totalItems", productsPage.getTotalElements());
-        model.addAttribute("pageSize", productsPage.getSize());
-
-        return "products_list";
-    }
-
-    //상품 추가 폼 페이지
-    @GetMapping("/new")
-    @Operation(summary = "상품 추가 폼", description = "상품 추가 폼 화면을 띄웁니다.")
-    public String createProductForm(Model model) {
-        List<Category> categories = categoryService.getAllCategories();
-        model.addAttribute("categories", categories);
-        return "form";
-    }
-
-    //상품 추가 데이터 응답
+    // 상품 추가 데이터 응답
     @PostMapping
-    @Operation(summary = "상품 추가", description = "상품 추가 폼에서 입력한 상품을 추가합니다.")
-    public String create(@Valid @ModelAttribute Product formProduct) {
+    public ResponseEntity<String> create(@Valid @RequestBody Product formProduct) {
         productService.addProduct(formProduct);
-        return "redirect:/products";
+        return ResponseEntity.status(HttpStatus.CREATED).body("Product added");
     }
 
-    //상품 단일 조회 기능
+    // 상품 단일 조회 기능
     @GetMapping("/{id}")
-    @Operation(summary = "단일 상품 조회", description = "단일 상품을 출력합니다.")
-    public String showOneProduct(@PathVariable("id") Long id, Model model) {
-        List<Product> products = new ArrayList<>();
-        products.add(productService.getProductById(id));
-        model.addAttribute("products", products);
-        return "products_list";
+    public ResponseEntity<Product> getProduct(@PathVariable("id") Long id) {
+        Product product = productService.getProductById(id);
+        return ResponseEntity.ok(product);
     }
 
-    //상품 검색 기능
-    @GetMapping("/search")
-    @Operation(summary = "상품 이름 검색", description = "검색어를 통해 해당 검색어를 포함한 이름의 상품들을 출력합니다.")
-    public String searchProduct(@Valid @ModelAttribute Product formProduct, Model model) {
-        List<Product> products = productService.searchProduct(formProduct.getName());
-        model.addAttribute("products", products);
-        return "products_list";
-    }
-
-    //상품 삭제 기능
+    // 상품 삭제 기능
     @DeleteMapping("/{id}")
-    @Operation(summary = "상품 제거", description = "해당 상품 제거")
-    public String deleteProduct(@PathVariable("id") Long id, Model model) {
-        Product product = productService.getProductById(id);
+    public ResponseEntity<Void> deleteProduct(@PathVariable("id") Long id) {
         productService.deleteProduct(id);
-        return "redirect:/products";
+        return ResponseEntity.ok().build();
     }
 
-    //상품 수정 폼 페이지
-    @GetMapping("/update/{id}")
-    @Operation(summary = "상품 수정 폼", description = "상품 수정 폼 화면을 띄웁니다.")
-    public String updateProductForm(@PathVariable("id") Long id, Model model) {
-        Product product = productService.getProductById(id);
-        model.addAttribute("product", product);
-
-        List<Category> categories = categoryService.getAllCategories();
-        model.addAttribute("categories", categories);
-        return "form";
-    }
-
-    //상품 수정 기능
+    // 상품 수정 기능
     @PutMapping("/{id}")
-    @Operation(summary = "상품 수정", description = "상품 수정 폼에서 입력한 상품을 수정합니다.")
-    public String updateProduct(@PathVariable("id") Long id,
-        @Valid @ModelAttribute Product updateProduct) {
+    public ResponseEntity<Product> updateProduct(@PathVariable("id") Long id,
+        @Valid @RequestBody Product updateProduct) {
         productService.updateProduct(id, updateProduct);
-        return "redirect:/products/" + id;
+        return ResponseEntity.ok().build();
     }
-
 }
